Data Pipeline Refactor: Move SemanticSegmentation to new DataModule API (#1038)
