+++ flash/core/data/input_transform.py
@@ -278,6 +278,40 @@ class InputTransform(Properties):
         on cpu for the predicting stage."""
         return self._identity
 
+    def serve_per_sample_transform(self) -> Callable:
+        """Defines the transform to be applied on a single sample on cpu for the serving stage.
+
+        The input data of the transform would have the following form:
+        {
+            DataKeys.INPUT: ...,
+            DataKeys.TARGET: ...,
+            DataKeys.METADATA: ...,
+        }
+
+        You would need to use :class:`flash.core.data.transforms.ApplyToKeys` as follows:
+
+        Example:
+
+            from flash.core.data.transforms import ApplyToKeys
+
+            class MyInputTransform(InputTransform):
+
+                def per_sample_transform(self) -> Callable:
+
+                    return ApplyToKeys("input", my_func)
+        """
+        return self._identity
+
+    def serve_input_per_sample_transform(self) -> Callable:
+        """Defines the transform to be applied on the value associated with the "input" key of each single sample
+        on cpu for the serving stage."""
+        return self._identity
+
+    def serve_target_per_sample_transform(self) -> Callable:
+        """Defines the transform to be applied on the value associated with the "target" key of each single sample
+        on cpu for the serving stage."""
+        return self._identity
+
     ##################################
     # PER SAMPLE TRANSFORM ON DEVICE #
     ##################################
@@ -579,6 +613,40 @@ class InputTransform(Properties):
         on cpu for the predicting stage."""
         return self._identity
 
+    def serve_per_batch_transform(self) -> Callable:
+        """Defines the transform to be applied on a batch of data on cpu for the serving stage.
+
+        The input data of the transform would have the following form:
+        {
+            DataKeys.INPUT: ...,
+            DataKeys.TARGET: ...,
+            DataKeys.METADATA: ...,
+        }
+
+        You would need to use :class:`flash.core.data.transforms.ApplyToKeys` as follows:
+
+        Example:
+
+            from flash.core.data.transforms import ApplyToKeys
+
+            class MyInputTransform(InputTransform):
+
+                def per_batch_transform(self) -> Callable:
+
+                    return ApplyToKeys("input", my_func)
+        """
+        return self._identity
+
+    def serve_input_per_batch_transform(self) -> Callable:
+        """Defines the transform to be applied on the value associated with the "input" key of each single sample
+        on cpu for the serving stage."""
+        return self._identity
+
+    def serve_target_per_batch_transform(self) -> Callable:
+        """Defines the transform to be applied on the value associated with the "target" key of each single sample
+        on cpu for the serving stage."""
+        return self._identity
+
     #################################
     # PER BATCH TRANSFORM ON DEVICE #
     #################################
@@ -747,6 +815,10 @@ class InputTransform(Properties):
 
     def predict_collate(self) -> Callable:
         """Defines the transform to be applied on a list of predicting sample to create a predicting batch."""
+        return default_collate
+
+    def serve_collate(self) -> Callable:
+        """Defines the transform to be applied on a list of serving sample to create a serving batch."""
         return default_collate
 
     def collate(self) -> Callable:
