+++ flash/image/segmentation/transforms.py
@@ -11,11 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Callable, Dict, Tuple
+from dataclasses import dataclass
+from typing import Any, Callable, Dict, Tuple
 
 import torch
 import torch.nn as nn
 
+from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import DataKeys
 from flash.core.utilities.imports import _KORNIA_AVAILABLE, _TORCHVISION_AVAILABLE
 
@@ -74,3 +76,45 @@ def predict_default_transforms(image_size: Tuple[int, int]) -> Dict[str, Callabl
         ),
         "collate": kornia_collate,
     }
+
+
+def remove_extra_dimensions(batch: Dict[str, Any]):
+    if isinstance(batch[DataKeys.INPUT], list):
+        assert len(batch[DataKeys.INPUT]) == 1
+        batch[DataKeys.INPUT] = batch[DataKeys.INPUT][0]
+    return batch
+
+
+@dataclass
+class SemanticSegmentationInputTransform(InputTransform):
+
+    image_size: Tuple[int, int] = (128, 128)
+
+    def train_per_sample_transform(self) -> Callable:
+        return ApplyToKeys(
+            [DataKeys.INPUT, DataKeys.TARGET],
+            KorniaParallelTransforms(
+                K.geometry.Resize(self.image_size, interpolation="nearest"), K.augmentation.RandomHorizontalFlip(p=0.5)
+            ),
+        )
+
+    def per_sample_transform(self) -> Callable:
+        return ApplyToKeys(
+            [DataKeys.INPUT, DataKeys.TARGET],
+            KorniaParallelTransforms(K.geometry.Resize(self.image_size, interpolation="nearest")),
+        )
+
+    def predict_input_per_sample_transform(self) -> Callable:
+        return K.geometry.Resize(self.image_size, interpolation="nearest")
+
+    def collate(self) -> Callable:
+        return kornia_collate
+
+    def target_per_batch_transform(self) -> Callable:
+        return prepare_target
+
+    def predict_per_batch_transform(self) -> Callable:
+        return remove_extra_dimensions
+
+    def serve_per_batch_transform(self) -> Callable:
+        return remove_extra_dimensions
